+++ crystal_toolkit/components/__init__.py
@@ -9,10 +9,10 @@ from crystal_toolkit.components.bandstructure import (
     BandstructureAndDosPanelComponent as BandstructureAndDosPanelComponent,
 )
 from crystal_toolkit.components.diffraction import (
-    TEMDiffractionComponent as TEMDiffractionComponent,
-)
-from crystal_toolkit.components.diffraction import (
     XRayDiffractionComponent as XRayDiffractionComponent,
+)
+from crystal_toolkit.components.diffraction_tem import (
+    TEMDiffractionComponent as TEMDiffractionComponent,
 )
 from crystal_toolkit.components.fermi_surface import (
     FermiSurfaceComponent as FermiSurfaceComponent,
